pyproject.toml
[tool.poetry]
name = "repl_python3_beta_miku"
version = "0.1.0"
description = ""
authors = ["DhruvaShaw <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
discord = "^1.7.3"
python-dotenv = "^0.18.0"
"discord.py" = "^1.7.3"
discord-pretty-help = "^1.3.1"
"statcord.py" = "^3.1.0"
discord-py-slash-command = "^2.0.4"
discord-ext-menus = "^1.1"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
